address review: drop BaseDomainError and relocate http errors to controllers
components must not be coupled to the http server. baseDomainError
exposed an httpStatus field that invited logic and adapter components
to extend it, baking transport semantics into the domain layer.

instead:

- delete src/errors.ts and the BaseDomainError class entirely.
- move InvalidRequestError and NotFoundError to
  src/controllers/errors.ts. they extend Error directly with no shared
  base. they are http-shaped errors meant to be thrown only by handlers
  to surface a specific status.
- replace the BaseDomainError + httpStatus mechanism in the error
  middleware with an explicit instanceof ladder against the http error
  classes. unknown errors continue to surface as 500.
- update all 13 handler files, controllers/utils.ts, and 2 spec files
  to import from src/controllers/errors.
- drop the transitional re-export from src/types.ts.

components that need to surface failures should define their own typed
errors in their own errors.ts file. each handler is responsible for
catching component errors and throwing the http error appropriate for
that endpoint — the same component error may map to different http
statuses across handlers.
