+++ content/src/controllers/handlers/create-entity-handler.ts
@@ -2,7 +2,8 @@ import { PostEntity200, PostEntity400 } from '@dcl/catalyst-api-specs/lib/client
 import { Field } from '@well-known-components/multipart-wrapper'
 import { AuthChain, Authenticator, AuthLink, EthAddress, Signature } from '@dcl/crypto'
 import { DeploymentContext, isInvalidDeployment, isSuccessfulDeployment } from '../../deployment-types'
-import { FormHandlerContextWithPath, InvalidRequestError } from '../../types'
+import { FormHandlerContextWithPath } from '../../types'
+import { InvalidRequestError } from '../errors'
 
 type ContentFile = {
   path?: string
