+++ content/src/controllers/handlers/get-available-content-handler.ts
@@ -1,5 +1,6 @@
 import { AvailableContent } from '@dcl/catalyst-api-specs/lib/client'
-import { HandlerContextWithPath, InvalidRequestError } from '../../types'
+import { HandlerContextWithPath } from '../../types'
+import { InvalidRequestError } from '../errors'
 import { qsGetArray, qsParser } from '../../logic/query-params'
 
 // Method: GET
