+++ content/src/controllers/handlers/get-deployments-handler.ts
@@ -2,7 +2,8 @@ import { EntityType } from '@dcl/schemas'
 import { asEnumValue, fromCamelCaseToSnakeCase } from '../utils'
 import { qsGetArray, qsGetBoolean, qsGetNumber, qsParser, toQueryParams } from '../../logic/query-params'
 import { Deployment, DeploymentBase, DeploymentOptions, SortingField, SortingOrder } from '../../deployment-types'
-import { DeploymentField, HandlerContextWithPath, InvalidRequestError, parseEntityType } from '../../types'
+import { DeploymentField, HandlerContextWithPath, parseEntityType } from '../../types'
+import { InvalidRequestError } from '../errors'
 import { getDeployments } from '../../logic/deployments'
 
 export const DEFAULT_FIELDS_ON_DEPLOYMENTS: DeploymentField[] = [
