+++ content/src/controllers/handlers/pointer-changes-handler.ts
@@ -2,7 +2,8 @@ import { PointerChanges } from '@dcl/catalyst-api-specs/lib/client'
 import { EntityType, PointerChangesSyncDeployment } from '@dcl/schemas'
 import { asEnumValue, fromCamelCaseToSnakeCase } from '../utils'
 import { qsGetArray, qsGetBoolean, qsGetNumber, qsParser, toQueryParams } from '../../logic/query-params'
-import { HandlerContextWithPath, InvalidRequestError, parseEntityType } from '../../types'
+import { HandlerContextWithPath, parseEntityType } from '../../types'
+import { InvalidRequestError } from '../errors'
 import { SortingField, SortingOrder } from '../../deployment-types'
 import { getPointerChanges, PointerChangesFilters } from '../../logic/deployments'
