+++ content/src/controllers/middlewares.ts
@@ -2,7 +2,7 @@ import { IHttpServerComponent, ILoggerComponent } from '@well-known-components/i
 import { AppComponents } from '../types'
 import { State } from '../adapters/synchronization-state'
 import { Error } from '@dcl/catalyst-api-specs/lib/client'
-import { BaseDomainError } from '../errors'
+import { InvalidRequestError, NotFoundError } from './errors'
 import { Middleware } from '@dcl/http-server/dist/middleware'
 
 export function preventExecutionIfBoostrapping({
@@ -27,16 +27,15 @@ export function preventExecutionIfBoostrapping({
 }
 
 function handleError(logger: ILoggerComponent.ILogger, error: any): { status: number; body: Error } {
-  // Typed domain errors carry their own canonical status. Subclasses of
-  // BaseDomainError (InvalidRequestError, NotFoundError, and any future
-  // per-component error class) get mapped automatically.
-  if (error instanceof BaseDomainError && typeof error.httpStatus === 'number') {
-    return {
-      status: error.httpStatus,
-      body: {
-        error: error.message
-      }
-    }
+  // Handlers throw HTTP-shaped errors (defined in `./errors.ts`) when they
+  // want to surface a specific status. Add a new branch when a new HTTP error
+  // class is introduced. Anything else is treated as an unexpected failure
+  // and surfaces as 500 to avoid leaking internal details.
+  if (error instanceof InvalidRequestError) {
+    return { status: 400, body: { error: error.message } }
+  }
+  if (error instanceof NotFoundError) {
+    return { status: 404, body: { error: error.message } }
   }
 
   logger.error(error)
