+++ content/src/types.ts
@@ -179,9 +179,6 @@ export type QueryParams = qs.ParsedQs
 
 export type AnyObject = Record<string, unknown>
 
-// transitional re-export; new code should import from './errors' directly
-export { BaseDomainError, InvalidRequestError, NotFoundError } from './errors'
-
 export enum DeploymentField {
   CONTENT = 'content',
   POINTERS = 'pointers',
