+++ content/test/unit/controllers/filter-by-urn-handler.spec.ts
@@ -1,6 +1,7 @@
 import { createMockedActiveEntitiesComponent } from '../../mocks/active-entities-component-mock'
 import { getEntitiesByCollectionPointerPrefixHandler } from '../../../src/controllers/handlers/filter-by-urn-handler'
-import { HandlerContextWithPath, InvalidRequestError } from '../../../src/types'
+import { HandlerContextWithPath } from '../../../src/types'
+import { InvalidRequestError } from '../../../src/controllers/errors'
 import { ActiveEntities } from '../../../src/logic/active-entities'
 import { createMockedEntity } from '../../mocks/entity-mock'
 import { Entity } from '@dcl/schemas'
